Move chart files to new folder
