+++ client/src/components/trends/chart/Area.jsx
