+++ client/src/components/trends/chart/Chart.jsx
