+++ client/src/components/trends/chart/chartUtils.js
